Handling of unsupported languages.
Whenever the daemons encounter an unsupported language they'll send the current
input to the last callback URL. It's up to the last callback URL to decide what
to do with the input document.

This change ensures that a single, unsupported language doesn't result in
documents being lost in the pipeline. Since any following callbacks are likely
not going to work either we can simply skip them, sending the data only to the
last one.